Update patch set 1
Patch Set 1:

/* getting out tinfoil hat */

I would be OK with this if it didn't override my disabling of it previously.  That is, after installing the most recent nightly build, CMStats was re-enabled.  Does that mean that if any updates are pushed out in the future (e.g. CM 7.0.1), that the user will have to know to disable it again?  Some would consider that malicious behavior.

Patch-set: 1