Update patch set 1
Patch Set 1:

Because I am a bit of a tin-foil hat person.   I appreciate the efforts of everyone contributing to CM, but I've seen projects fall down that slippery slope of making minor changes that seemed innocuous at first, but led to onerous changes that no one likes. Installing any service that sends info every time I boot the phone definitely seems like it is leading down that path.

If limited to sending in a notification when first installed or updated, I would be OK with.  Every time I boot the phone?  I'll pass on that.

Patch-set: 1