Update patch set 1
Patch Set 1:

That's basically the only reason, yes.  Besides the stats that you can see when you preview the statistics, we don't collect any data other than first and last checkin.  If the last checkin is > 30 days from the current time, the device gets dropped from the DB and all the info swept clean. Any app with the READ_PHONE_STATE permission can get the deviceID that we do, but we even hash it before uploading it to the servers, protecting your identity further

Patch-set: 1